// ClientApp/src/Types/IProductClasses.ts
export default interface IProductClasses {
  classes_fromPlace: string;
  classes_fromPlaceName: string;
  classes_toPlace: string;
  classes_toPlaceName: string;

  classes_activatedBy: string;
  classes_activatedDate: string;
  classes_classID: string;
  classes_code: string;
  classes_createdBy: string;
  classes_createdDate: string;
  classes_deactivatedBy: string;
  classes_deactivatedDate: string;
  classes_descriptions: string;
  classes_editedBy: string;
  classes_editedDate: string;
  classes_isActive: boolean;
  classes_name: string;
  classes_price: number;
  classes_productID: string;
  classes_ticketsRangedFrom: number;
  classes_ticketsRangedTo: number;
  classes_totalNumberOfTickets: number;
  products_branchID: string;
  products_code: string;
  products_descriptions: string;
  products_fromPlace: string;
  products_isActive: boolean;
  products_name: string;
  products_productID: string;
  products_subcategoryID: number;
  products_toPlace: string;
}